+++ Common/src/main/java/com/tristankechlo/toolleveling/client/ToolLevelingTableScreen.java
@@ -14,18 +14,23 @@ import net.minecraft.network.chat.MutableComponent;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.util.Tuple;
 import net.minecraft.world.entity.player.Inventory;
-import net.minecraft.world.item.enchantment.Enchantment;
 
 import java.util.List;
+import java.util.stream.Collectors;
 
 public class ToolLevelingTableScreen extends AbstractContainerScreen<ToolLevelingTableMenu> {
 
     private static final ResourceLocation GUI_TEXTURE = new ResourceLocation(ToolLeveling.MOD_ID, "textures/gui/tool_leveling_table.png");
+    private static final Tooltip TOOLTIP_PERCENTAGES = Tooltip.create(make(".tooltip.percentages"));
+    private static final Tooltip TOOLTIP_HELP = Tooltip.create(make(".tooltip.help"));
+    private static final Component PERCENTAGES_TITLE = make(".percentages.title").withStyle(ChatFormatting.GRAY);
     private static boolean shouldRenderPercentages = false;
-    List<Tuple<Enchantment, Float>> percentages;
-    private float successChance;
+    private static boolean shouldRenderHelp = false;
+    private final InfoFieldRenderer percentagesInfoField = new InfoFieldRenderer(0xFF212121, 0xFF7C0053, 0xFFD82FA0);
+    private final InfoFieldRenderer successChanceField = new InfoFieldRenderer(0xFF212121, 0xFF3B51BF, 0xFF4F80FF);
     private byte ticksSinceUpdate = 0;
     private Button infoButton;
+    private Button helpButton;
 
     public ToolLevelingTableScreen(ToolLevelingTableMenu container, Inventory inv, Component name) {
         super(container, inv, name);
@@ -38,14 +43,14 @@ public class ToolLevelingTableScreen extends AbstractContainerScreen<ToolLevelin
     @Override
     protected void init() {
         super.init();
-        this.percentages = this.getMenu().getPercentages();
-        this.successChance = Util.getSuccessChance(this.getMenu());
-        Button.OnPress action = (button) -> {
-            shouldRenderPercentages = !shouldRenderPercentages;
-        };
-        this.infoButton = this.addRenderableWidget(new Button.Builder(Component.literal("?"), action)
-                .pos(this.leftPos + this.imageWidth - 20, this.topPos + 4).size(16, 16)
-                .tooltip(Tooltip.create(Component.literal("This is a test"))).build());
+        this.infoButton = this.addRenderableWidget(new Button.Builder(Component.literal("%"), (b) -> shouldRenderPercentages = !shouldRenderPercentages)
+                .pos(this.leftPos + this.imageWidth - 34, this.topPos + 94).size(14, 14)
+                .tooltip(TOOLTIP_PERCENTAGES).build());
+        this.helpButton = this.addRenderableWidget(new Button.Builder(Component.literal("?"), (b) -> shouldRenderHelp = !shouldRenderHelp)
+                .pos(this.leftPos + this.imageWidth - 18, this.topPos + 94).size(14, 14)
+                .tooltip(TOOLTIP_HELP).build());
+
+        this.percentagesInfoField.setSpaceAfterTitle(true);
     }
 
     @Override
@@ -57,9 +62,18 @@ public class ToolLevelingTableScreen extends AbstractContainerScreen<ToolLevelin
         this.ticksSinceUpdate++;
         if (this.ticksSinceUpdate % 5 == 0) {
             this.ticksSinceUpdate = 0;
-            percentages = this.getMenu().getPercentages();
+
+            // update percentages
+            var percentages = this.getMenu().getPercentages();
             percentages.sort((o1, o2) -> Float.compare(o2.getB(), o1.getB()));
-            this.successChance = Util.getSuccessChance(this.getMenu());
+            var components = percentages.stream().map(ToolLevelingTableScreen::createComponent).collect(Collectors.toList());
+            components.add(0, PERCENTAGES_TITLE);
+            this.percentagesInfoField.setLines(components);
+
+            // update success chance
+            float successChance = Util.getSuccessChance(this.getMenu());
+            Component chanceText = createComponent(new Tuple<>("screen.toolleveling.tool_leveling_table.success_chance", successChance));
+            this.successChanceField.setLines(List.of(chanceText));
         }
     }
 
@@ -69,55 +83,54 @@ public class ToolLevelingTableScreen extends AbstractContainerScreen<ToolLevelin
         super.render(poseStack, mouseX, mouseY, partialTicks);
         this.renderTooltip(poseStack, mouseX, mouseY); // render tooltips if hovered
 
-        if (!shouldRenderPercentages || this.percentages.isEmpty()) {
-            return;
+        if (shouldRenderPercentages) {
+            int x = this.leftPos + this.imageWidth;
+            int y = this.topPos;
+            int fieldWidth = this.calcFieldWidth();
+            if (this.getMenu().hasAnyBooks()) {
+                this.percentagesInfoField.render(poseStack, this.font, x, y, fieldWidth);
+                y += this.percentagesInfoField.calcHeight() + 1;
+            }
+            this.successChanceField.render(poseStack, this.font, x, y, fieldWidth);
         }
+    }
 
-        int x = this.leftPos + this.imageWidth + 7;
-        int y = this.topPos + 7;
-        for (int i = 0; i < this.percentages.size(); i++) {
-            MutableComponent text = createComponent(this.percentages.get(i));
-            drawString(poseStack, this.font, text, x, y + (i * 11), 0xFFFFFF);
-        }
-
-        MutableComponent text = Component.literal("Success Chance: ").withStyle(ChatFormatting.GRAY);
-        text.append(Component.literal(this.successChance + "%").withStyle(ChatFormatting.WHITE));
-        drawString(poseStack, this.font, text, x, y + (this.percentages.size() * 11) + 4, 0xFFFFFF);
+    private int calcFieldWidth() {
+        int width1 = this.percentagesInfoField.calcWidth(this.font);
+        int width2 = this.successChanceField.calcWidth(this.font);
+        int targetWidth = Math.max(width1, width2); // get the widest field
+        int widthFree = this.width - this.leftPos - this.imageWidth - 3; // get the free space on the right side
+        targetWidth = Math.min(targetWidth, widthFree); // limit width to screen width
+        return targetWidth;
     }
 
     @Override
     protected void renderBg(PoseStack poseStack, float partialTicks, int mX, int my) {
         RenderSystem.setShaderTexture(0, GUI_TEXTURE);
         blit(poseStack, this.leftPos, this.topPos, 0, 0, this.imageWidth, this.imageHeight);
-
-        if (!shouldRenderPercentages || this.percentages.isEmpty()) {
-            return;
-        }
-
-        int x = this.leftPos + this.imageWidth;
-        int y = this.topPos + 6;
-        blit(poseStack, x, y - 6, 0, 224, 120, 6);
-        for (int i = 0; i < this.percentages.size(); i++) {
-            blit(poseStack, x, y + (i * 11), 0, 228, 120, 12);
-        }
-        blit(poseStack, x, y + (this.percentages.size() * 11), 0, 252, 120, 4);
     }
 
-    private static MutableComponent createComponent(Tuple<Enchantment, Float> tuple) {
-        Enchantment enchantment = tuple.getA();
+    private static Component createComponent(Tuple<String, Float> tuple) {
         float percentage = tuple.getB();
-        percentage = Math.round(percentage * 1000F) / 10F; // round to 1 decimal place
-        MutableComponent text = Component.translatable(enchantment.getDescriptionId()).withStyle(ChatFormatting.GRAY);
+        percentage = Math.round(percentage * 10000F) / 100F; // round to 2 decimal place
+        MutableComponent text = Component.translatable(tuple.getA()).withStyle(ChatFormatting.GRAY);
         text.append(Component.literal(" " + percentage + "%").withStyle(ChatFormatting.GREEN));
         return text;
     }
 
     @Override
-    public boolean mouseClicked(double $$0, double $$1, int $$2) {
-        if (!this.infoButton.isMouseOver($$0, $$1)) {
+    public boolean mouseClicked(double mouseX, double mouseY, int button) {
+        if (!this.infoButton.isMouseOver(mouseX, mouseY)) {
             this.infoButton.setFocused(false); // unfocus button if clicked outside of it
         }
-        return super.mouseClicked($$0, $$1, $$2);
+        if (!this.helpButton.isMouseOver(mouseX, mouseY)) {
+            this.helpButton.setFocused(false); // unfocus button if clicked outside of it
+        }
+        return super.mouseClicked(mouseX, mouseY, button);
+    }
+
+    private static MutableComponent make(String string) {
+        return Component.translatable("screen.toolleveling.tool_leveling_table" + string);
     }
 
 }