+++ Common/src/main/java/com/tristankechlo/toolleveling/config/CommandConfig.java
@@ -1,0 +1,10 @@
+package com.tristankechlo.toolleveling.config;
+
+import java.util.function.Supplier;
+
+public final class CommandConfig {
+
+    public static final Supplier<Boolean> allowWrongEnchantments = () -> true;
+    public static final Supplier<Boolean> allowIncompatibleEnchantments = () -> true;
+
+}