+++ Common/src/main/java/com/tristankechlo/toolleveling/config/ToolLevelingConfig.java
@@ -1,14 +1,27 @@
 package com.tristankechlo.toolleveling.config;
 
 import net.minecraft.world.item.Item;
+import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.Items;
 
 import java.util.function.Supplier;
 
 public final class ToolLevelingConfig {
 
-    public static final Supplier<Float> MIN_SUCCESS_CHANCE = () -> 0.5F;
-    public static final Supplier<Float> MAX_SUCCESS_CHANCE = () -> 1.0F;
-    public static final Supplier<Item> PAYMENT_ITEM = () -> Items.LAPIS_LAZULI;
+    public static final Supplier<Float> minSuccessChance = () -> 75.0F;
+    public static final Supplier<Float> maxSuccessChance = () -> 95.0F;
+    private static final Supplier<Item> paymentItem = () -> Items.LAPIS_LAZULI;
+
+    public static boolean isPaymentItem(ItemStack stack) {
+        return stack.is(paymentItem.get());
+    }
+
+    public static boolean isBonusItemHigherLevel(ItemStack stack) {
+        return stack.is(Items.ENCHANTED_GOLDEN_APPLE);
+    }
+
+    public static boolean isBonusItemExtraEnchantment(ItemStack stack) {
+        return stack.is(Items.NETHER_STAR);
+    }
 
 }